requirements.txt
bandit==1.7.0
cloudpickle==2.0.0
cycler==0.10.0
EasyProcess==0.3
gitdb==4.0.7
GitPython==3.1.24
gym==0.20.0
kiwisolver==1.3.2
matplotlib==3.4.3
numpy==1.21.2
pandas==1.3.3
pbr==5.6.0
Pillow==8.3.2
pyglet==1.5.21
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
PyVirtualDisplay==2.2
PyYAML==5.4.1
scipy==1.7.1
seaborn==0.11.2
six==1.16.0
smmap==4.0.0
stevedore==3.4.0
torch==1.9.0
tqdm==4.62.3
typing-extensions==3.10.0.2
